Reply migration forein key fixed

+++ database/migrations/2019_10_05_203055_create_replies_table.php
@@ -17,8 +17,7 @@ class CreateRepliesTable extends Migration
             $table->bigIncrements('id');
             $table->text('body');
 
-            $table->integer('question_id')->unsigned();
-
+            $table->unsignedBigInteger('question_id')->unsigned();
             $table->integer('user_id')->unsigned();
 
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
